refactor(test): consolidate AA processor IT tests
Create a base class AAChannelProcessorBaseIT with shared fields (archiveProperty, activeProperty, inactiveProperty),
@MockitoBean/@Autowired declarations, @BeforeEach primeCache(), and the
paramableAAChannelProcessorTest helper.
