+++ src/test/java/org/phoebus/channelfinder/processors/aa/AAChannelProcessorTagPauseIT.java
@@ -1,44 +1,23 @@
 package org.phoebus.channelfinder.processors.aa;
 
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.when;
-import static org.phoebus.channelfinder.processors.aa.AAChannelProcessorIT.archiveProperty;
-import static org.phoebus.channelfinder.processors.aa.AAChannelProcessorIT.inactiveProperty;
-import static org.phoebus.channelfinder.processors.aa.AAChannelProcessorIT.paramableAAChannelProcessorTest;
-
 import java.util.List;
 import java.util.stream.Stream;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.phoebus.channelfinder.configuration.AAChannelProcessor;
 import org.phoebus.channelfinder.entity.Channel;
-import org.phoebus.channelfinder.service.external.ArchiverService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import tools.jackson.core.JacksonException;
 
 @WebMvcTest(AAChannelProcessor.class)
 @TestPropertySource(
     locations = "classpath:application_aa_proc_test.properties",
     properties = "aa.auto_pause=archive")
-class AAChannelProcessorTagPauseIT {
-
-  @Autowired AAChannelProcessor aaChannelProcessor;
-
-  @MockitoBean ArchiverService archiverService;
-
-  @BeforeEach
-  void primeCache() {
-    when(archiverService.getAAPolicies(anyString())).thenReturn(List.of("policy"));
-    aaChannelProcessor.scheduledPolicyRefresh();
-  }
+class AAChannelProcessorTagPauseIT extends AAChannelProcessorBaseIT {
 
   private static Stream<Arguments> processNoPauseSource() {
-
     return Stream.of(
         Arguments.of(
             new Channel(
@@ -61,7 +40,6 @@ class AAChannelProcessorTagPauseIT {
   void testProcessNotArchivedActive(
       Channel channel, String archiveStatus, String archiverEndpoint, String submissionBody)
       throws JacksonException {
-    paramableAAChannelProcessorTest(
-        archiverService, aaChannelProcessor, List.of(channel), archiveStatus, archiverEndpoint);
+    paramableAAChannelProcessorTest(List.of(channel), archiveStatus, archiverEndpoint);
   }
 }
